requirements.txt
O365>=2.0.34
openai>=1.58.1
python-dotenv>=1.0.1
anthropic>=0.42.0
customtkinter
requests>=2.31.0